DoFOutputOperator: consolidate things to a single header.
There are only two functions in the .templates.h file: let's just move
everything over.
