+++ source/numerics/dof_output_operator.cc
@@ -11,9 +11,6 @@
 // -----------------------------------------------------------------------------
 
 
-#include <deal.II/base/logstream.h>
-#include <deal.II/base/parameter_handler.h>
-
 #include <deal.II/lac/block_vector.h>
 #include <deal.II/lac/la_parallel_block_vector.h>
 #include <deal.II/lac/la_parallel_vector.h>
@@ -22,9 +19,8 @@
 #include <deal.II/lac/trilinos_parallel_block_vector.h>
 #include <deal.II/lac/trilinos_vector.h>
 #include <deal.II/lac/vector.h>
-#include <deal.II/lac/vector_memory.h>
 
-#include <deal.II/numerics/dof_output_operator.templates.h>
+#include <deal.II/numerics/dof_output_operator.h>
 
 
 DEAL_II_NAMESPACE_OPEN
